P0 — stop the untruths
Four things the product reported that were not so.

**Upload never completed.** The worker wrote state "done"; the route asked
whether it was "finished". Two vocabularies for one thing, neither wrong alone.
The screen polled forever, showed real progress the whole time, and never said
it had ended. There is one State class now and one TERMINAL set.

**No test has ever gated a deployment.** CI triggered on pull requests and on
pushes to branches other than main. Everything here reaches production by a push
to main, and deploy.yml had no test step — so ci.yml has run zero times. The job
was even named "Agents & API" while every step ran in agents/. CI now runs
agents, API, web typecheck and web build, and deploy calls it as a reusable
workflow and needs it.

Which uncovered why the API tests could never have run in CI: api imports
trimbin_agents and never declared it. It worked because one virtualenv had it
installed editable. Declared now as a path source.

And the API had never been linted anywhere — 44 errors. B008, N818 and the
ambiguous-unicode rules are ignored with reasons; the rest are fixed.

**The margin rule had four implementations.** The dot in the tree, the reason in
the queue, the unconfirmed flag on the assembly, and the field a comparison
returns. Same threshold, different logic — the assembly ignored the circled
take, the comparison ignored whether anyone had looked. A scene could report
"all settled" while the dashboard listed three of its shots as waiting. One rule
in services/assessment.py, and an exhaustive test that status and reason can
never disagree again.

**The vector index was never on the table.** It is declared inside CREATE TABLE
IF NOT EXISTS in 001_clips.sql and was added to that file after the table
existed, so every migration since found the table present and did nothing:

  SELECT name FROM system.data_skipping_indices WHERE table = 'clips'
  -> idx_description, idx_duration

Added by ALTER, and the verifier counts indexes now instead of only tables.

Also correcting something I reported wrongly: 2 clips carry an embedding, not
15. Of 306,230 rows. The backfill is P3, where G7 already lives.

291 API tests, lint clean.

+++ api/app/services/exports.py
@@ -46,9 +46,9 @@ def timecode(seconds: float, fps: float = DEFAULT_FPS) -> str:
     """
     if seconds < 0:
         seconds = 0.0
-    total_frames = int(round(seconds * fps))
-    frames = total_frames % int(round(fps))
-    total_seconds = total_frames // int(round(fps))
+    total_frames = round(seconds * fps)
+    frames = total_frames % round(fps)
+    total_seconds = total_frames // round(fps)
     return (
         f"{total_seconds // 3600:02d}:"
         f"{(total_seconds % 3600) // 60:02d}:"
